+++ lingting-core/src/main/kotlin/live/lingting/framework/money/MoneyResultCode.kt
@@ -3,9 +3,10 @@ package live.lingting.framework.money
 import live.lingting.framework.api.ResultCode
 
 /**
+ * 202201
  * @author lingting 2023-05-07 18:02
  */
-enum class MoneyResultCode(override val code: Int, override val message: String) : ResultCode {
+enum class MoneyResultCode(override val code: Long, override val message: String) : ResultCode {
     /**
      * 金额值异常!
      */
